:art: better asynchronous controllers methods

+++ controllers/todosController.js
@@ -2,42 +2,64 @@ const Todo = require("../models/TODO")
 
 module.exports = {
   getAllTodos: async (req, res) => {
-    Todo.find()
-      .then((todos) => res.status(200).send(todos))
-      .catch((err) => res.status(500).send(err))
+    try {
+      const todos = await Todo.find()
+      if (!todos) return res.status(404).send("No todos found")
+      return res.status(200).send(todos)
+    } catch (err) {
+      return res.status(500).send(err)
+    }
   },
   getTodo: async (req, res) => {
     const { id } = req.params
-    // TODO : vérifier que le todo existe
-    Todo.findById(id)
-      .then((todo) => res.status(200).send(todo))
-      .catch((err) => res.status(500).send(err))
+    if (!id.match(/^[0-9a-fA-F]{24}$/))
+      return res.status(400).send("Not a valid id format")
+    try {
+      const todo = await Todo.findById(id)
+      if (!todo) return res.status(404).send("Todo not found")
+      return res.status(200).send(todo)
+    } catch (err) {
+      return res.status(500).send(err)
+    }
   },
   addTodo: async (req, res) => {
     const { title, description, done } = req.body
-    const todo = new Todo({
+    const newTodo = new Todo({
       title,
       description,
       done,
     })
-    todo
-      .save()
-      .then((todo) => res.status(201).send(todo))
-      .catch((err) => res.status(500).send(err))
+    try {
+      const todo = await newTodo.save()
+      return res.status(201).send(todo)
+    } catch (err) {
+      res.status(500).send(err)
+    }
   },
   updateTodo: async (req, res) => {
     const { id } = req.params
     const { title, description, done } = req.body
-    //TODO : vérifier que le todo existe
-    Todo.findByIdAndUpdate(id, { title, description, done })
-      .then((todo) => res.status(200).send(todo))
-      .catch((err) => res.status(500).send(err))
+    if (!id.match(/^[0-9a-fA-F]{24}$/))
+      return res.status(400).send("Not a valid id format")
+    try {
+      const todo = await Todo.findById(id)
+      if (!todo) return res.status(404).send("Todo not found")
+      await Todo.findByIdAndUpdate(id, { title, description, done })
+      return res.status(200).send(todo)
+    } catch (err) {
+      res.status(500).send(err)
+    }
   },
   deleteTodo: async (req, res) => {
     const { id } = req.params
-    //TODO : vérifier que le todo existe
-    Todo.findByIdAndDelete(id)
-      .then((todo) => res.status(200).send(todo))
-      .catch((err) => res.status(500).send(err))
+    if (!id.match(/^[0-9a-fA-F]{24}$/))
+      return res.status(400).send("Not a valid id format")
+    try {
+      const todo = await Todo.findByIdAndDelete(id)
+      if (!todo) return res.status(404).send("Todo not found")
+      return res.status(200).send(todo)
+    } catch (err) {
+      return res.status(500).send(err)
+    }
   },
 }
